Updated site characterstics management , refactor for reuse
