+++ opensrp-anc/src/main/AndroidManifest.xml
@@ -56,6 +56,7 @@
             android:launchMode="singleTask"
             android:screenOrientation="portrait"
             android:theme="@style/AncAppTheme.NoActionBar" />
+
         <activity
             android:name=".activity.ContactActivity"
             android:configChanges="keyboardHidden|orientation|screenSize"
