+++ opensrp-anc/src/main/java/org/smartregister/anc/activity/SiteCharacteristicsActivity.java
@@ -1,17 +1,15 @@
 package org.smartregister.anc.activity;
 
-import android.os.Bundle;
-
-import org.smartregister.anc.R;
+import org.smartregister.anc.contract.BaseCharacteristicsContract;
+import org.smartregister.anc.presenter.SiteCharacteristicsViewPresenter;
 
 /**
  * Created by ndegwamartin on 30/08/2018.
  */
-public class SiteCharacteristicsActivity extends BaseActivity {
+public class SiteCharacteristicsActivity extends BaseCharacteristicsActivity {
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_characteristics);
+    protected BaseCharacteristicsContract.BasePresenter getPresenter() {
+        return new SiteCharacteristicsViewPresenter(this);
     }
 }
