+++ opensrp-anc/src/main/java/org/smartregister/anc/listener/NavigationItemListener.java
@@ -9,6 +9,7 @@ import android.view.View;
 import org.smartregister.anc.R;
 import org.smartregister.anc.activity.HomeRegisterActivity;
 import org.smartregister.anc.activity.PopulationCharacteristicsActivity;
+import org.smartregister.anc.activity.SiteCharacteristicsActivity;
 import org.smartregister.anc.application.AncApplication;
 
 /**
@@ -35,7 +36,8 @@ public class NavigationItemListener implements View.OnClickListener {
             // TODO go to counseling resources page
         } else if (id == R.id.site_characteristics) {
 
-            // TODO go to site characteristics page
+            context.startActivity(new Intent(context, SiteCharacteristicsActivity.class));
+
         } else if (id == R.id.population_characteristics) {
 
             context.startActivity(new Intent(context, PopulationCharacteristicsActivity.class));
