+++ opensrp-anc/src/main/java/org/smartregister/anc/task/FetchSiteCharacteristicsTask.java
@@ -2,6 +2,7 @@ package org.smartregister.anc.task;
 
 import android.os.AsyncTask;
 
+import org.smartregister.anc.contract.BaseCharacteristicsContract;
 import org.smartregister.anc.contract.PopulationCharacteristicsContract;
 import org.smartregister.anc.domain.Characteristic;
 import org.smartregister.anc.helper.CharacteristicsHelper;
@@ -11,17 +12,17 @@ import java.util.List;
 /**
  * Created by ndegwamartin on 28/08/2018.
  */
-public class FetchPopulationCharacteristicsTask extends AsyncTask<Void, Void, List<Characteristic>> {
+public class FetchSiteCharacteristicsTask extends AsyncTask<Void, Void, List<Characteristic>> {
 
-    private PopulationCharacteristicsContract.Presenter presenter;
+    private BaseCharacteristicsContract.BasePresenter presenter;
 
-    public FetchPopulationCharacteristicsTask(PopulationCharacteristicsContract.Presenter presenter) {
+    public FetchSiteCharacteristicsTask(PopulationCharacteristicsContract.Presenter presenter) {
         this.presenter = presenter;
     }
 
     @Override
     protected List<Characteristic> doInBackground(final Void... params) {
-        CharacteristicsHelper helper = new CharacteristicsHelper("population_characteristics");
+        CharacteristicsHelper helper = new CharacteristicsHelper("site_characteristics");
         return helper.getPopulationCharacteristics();
     }
