+++ opensrp-anc/src/main/res/layout/activity_characteristics.xml
@@ -4,6 +4,8 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
+    <include layout="@layout/toolbar_population_characteristics" />
+
     <android.support.v7.widget.RecyclerView
         android:id="@+id/population_characteristics"
         android:layout_width="match_parent"
