--- opensrp-anc/src/main/res/layout/activity_population_characteristics_row.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:paddingBottom="@dimen/population_characteristics_item_spacing"
-    android:paddingLeft="@dimen/population_characteristics_drawable_padding"
-    android:paddingRight="@dimen/population_characteristics_drawable_padding"
-    android:paddingTop="@dimen/population_characteristics_item_spacing">
-
-    <TextView
-        android:id="@+id/label"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:layout_centerVertical="true"
-        android:layout_marginLeft="10dp"
-        android:layout_toLeftOf="@id/value"
-        android:lineSpacingExtra="@dimen/population_characteristics_line_spacing"
-        android:text="Undernourished 20% or higher"
-        android:textAppearance="@style/TextAppearance.AppCompat.Medium"
-        android:textColor="@color/text_black" />
-
-    <TextView
-        android:id="@+id/value"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerVertical="true"
-        android:layout_toLeftOf="@id/info"
-        android:padding="@dimen/population_characteristics_padding"
-        android:text="No"
-        android:textAppearance="@style/TextAppearance.AppCompat.Medium"
-        android:textColor="@color/text_black" />
-
-    <ImageView
-        android:id="@+id/info"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_centerVertical="true"
-        android:layout_gravity="center_vertical"
-        android:padding="@dimen/population_characteristics_padding"
-        android:src="@drawable/ic_info" />
-
-</RelativeLayout>
